Remove retries on aborts and persists
For the moment this simplification is an improvement,
but technically COMMIT PREPARED should retry until it is successful
in order to guarantee two-phase commit properties. State needs
to be added anyway, in order to assure that if the transaction
manager goes down (which is the Idris application in this instance),
that the transaction can be resumed.